refactor: user-actions test
